Refactory ClassLoaderThreadFactory to wildfly-clustering-service.
